chore: fix data races pointed by the race detector
